[ADD] added entire VGA average feature

+++ src/main/java/com/example/vga_price_tracker/dto/VgaPriceDTO.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 @Getter
 public class VgaPriceDTO {   // 그래픽카드 가격 DTO
     private final String vgaName;
-    private final float vgaPrice;
+    private final double vgaPrice;
     private final String date;
 
     @Builder
-    public VgaPriceDTO(String vgaName, float vgaPrice, String date) {
+    public VgaPriceDTO(String vgaName, double vgaPrice, String date) {
         this.vgaName = vgaName;
         this.vgaPrice = vgaPrice;
         this.date = date;
